+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * (theimo1221) Clear Persisting from old mixed persisting commands
 * (theimo1221) Make battery readonly to outside
 * (theimo1221) Persist Battery Device Data
+* (theimo1221) Add iDisposable for Objects with intervals
+* (theimo1221) Persist Zigbee Device Data
 
 ## 1.1.48 (2022-10-15)